+++ CMakeLists.txt
@@ -9,6 +9,4 @@ endif()
 
 add_subdirectory(3rdparty EXCLUDE_FROM_ALL)
 add_subdirectory(code)
-add_subdirectory(code_v3)
-add_subdirectory(code_v4)
 add_subdirectory(example)
