+++ example/CMakeLists.txt
@@ -1,29 +1,29 @@
-set(TargetName elements)
-add_executable(${TargetName} elements.cpp)
-target_compile_features(${TargetName} PRIVATE cxx_std_20)
-target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
-target_link_libraries(${TargetName} PRIVATE xxx::xxx)
-
-set(TargetName stats)
-add_executable(${TargetName} stats.cpp)
-target_compile_features(${TargetName} PRIVATE cxx_std_20)
-target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
-target_link_libraries(${TargetName} PRIVATE xxx::xxx)
-
-set(TargetName test)
-add_executable(${TargetName} test.cpp)
-target_compile_features(${TargetName} PRIVATE cxx_std_23)
-target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
-target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
-
-set(TargetName test2)
-add_executable(${TargetName} test2.cpp)
-target_compile_features(${TargetName} PRIVATE cxx_std_23)
-target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
-target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
-
-set(TargetName test3)
-add_executable(${TargetName} test3.cpp)
-target_compile_features(${TargetName} PRIVATE cxx_std_23)
-target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
-target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
+# set(TargetName elements)
+# add_executable(${TargetName} elements.cpp)
+# target_compile_features(${TargetName} PRIVATE cxx_std_20)
+# target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
+# target_link_libraries(${TargetName} PRIVATE xxx::xxx)
+#
+# set(TargetName stats)
+# add_executable(${TargetName} stats.cpp)
+# target_compile_features(${TargetName} PRIVATE cxx_std_20)
+# target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
+# target_link_libraries(${TargetName} PRIVATE xxx::xxx)
+#
+# set(TargetName test)
+# add_executable(${TargetName} test.cpp)
+# target_compile_features(${TargetName} PRIVATE cxx_std_23)
+# target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
+# target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
+#
+# set(TargetName test2)
+# add_executable(${TargetName} test2.cpp)
+# target_compile_features(${TargetName} PRIVATE cxx_std_23)
+# target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
+# target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
+#
+# set(TargetName test3)
+# add_executable(${TargetName} test3.cpp)
+# target_compile_features(${TargetName} PRIVATE cxx_std_23)
+# target_compile_options(${TargetName} PRIVATE -Wall -Wextra -g)
+# target_link_libraries(${TargetName} PRIVATE xxx::xxx_v3)
